# api/openapi/swagger.yaml
components:
  schemas:
    errs.Error:
      properties:
        code:
          $ref: '#/components/schemas/errs.ErrorCode'
        message:
          type: string
        params:
          items:
            $ref: '#/components/schemas/errs.Param'
          type: array
          uniqueItems: false
      type: object
    errs.ErrorCode:
      type: integer
      x-enum-varnames:
      - ErrorCodeOK
      - ErrorCodeCanceled
      - ErrorCodeUnknown
      - ErrorCodeInvalidArgument
      - ErrorCodeDeadlineExceeded
      - ErrorCodeNotFound
      - ErrorCodeAlreadyExists
      - ErrorCodePermissionDenied
      - ErrorCodeResourceExhausted
      - ErrorCodeFailedPrecondition
      - ErrorCodeAborted
      - ErrorCodeOutOfRange
      - ErrorCodeUnimplemented
      - ErrorCodeInternal
      - ErrorCodeUnavailable
      - ErrorCodeDataLoss
      - ErrorCodeUnauthenticated
      - ErrorCodeClosedRequest
    errs.Param:
      properties:
        key:
          type: string
        value:
          type: string
      type: object
...
    handlers.TagCreateDTO:
      properties:
        post_id:
          type: string
        value:
          type: string
      type: object
    handlers.TagDTO:
      properties:
        created_at:
          type: string
        id:
          type: string
        post_id:
          type: string
        updated_at:
          type: string
        value:
          type: string
      type: object
    handlers.TagListDTO:
      properties:
        count:
          type: integer
        items:
          items:
            $ref: '#/components/schemas/handlers.TagDTO'
          type: array
          uniqueItems: false
      type: object
    handlers.TagUpdateDTO:
      properties:
        id:
          type: string
        post_id:
          type: string
        value:
          type: string
      type: object
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  title: example
  version: 0.0.0
openapi: 3.1.0
paths:
...
  /api/v1/posts/tags/:
    get:
      parameters:
      - in: query
        name: page_size
        schema:
          type: integer
      - in: query
        name: page_number
        schema:
          type: integer
      - in: query
        name: order_by
        schema:
          items:
            type: string
          type: array
          uniqueItems: false
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TagListDTO'
          description: Filtered list of tags
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: List of tags
      tags:
      - tag
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.TagCreateDTO'
        description: Create tag request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TagDTO'
          description: Created tag
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Create tag
      tags:
      - tag
  /api/v1/posts/tags/{id}:
    delete:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "204":
          description: No content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Delete tag by id
      tags:
      - tag
    get:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TagDTO'
          description: Requested tag
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Get tag by id
      tags:
      - tag
    patch:
      parameters:
      - description: UUID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.TagUpdateDTO'
        description: Update tag request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.TagDTO'
          description: Updated tag
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Invalid request body or validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errs.Error'
          description: Internal server error
      security:
      - BearerAuth: []
      summary: Update tag
      tags:
      - tag
servers:
- url: http://127.0.0.1:8000/